+++ internal/bundler/linker.go
@@ -3783,27 +3783,27 @@ func (repr *chunkReprJS) generate(c *linkerContext, chunk *chunkInfo) func(gener
 		}
 
 		// Start the metadata
-		jMeta := js_printer.Joiner{}
-		if c.options.AbsMetadataFile != "" {
+		sbMeta := strings.Builder{}
+		if c.options.NeedsMetafile {
 			// Print imports
 			isFirstMeta := true
-			jMeta.AddString("{\n      \"imports\": [")
+			sbMeta.WriteString("{\n      \"imports\": [")
 			for i, importAbsPath := range continueData.crossChunkAbsPaths {
 				if isFirstMeta {
 					isFirstMeta = false
 				} else {
-					jMeta.AddString(",")
+					sbMeta.WriteString(",")
 				}
-				jMeta.AddString(fmt.Sprintf("\n        {\n          \"path\": %s,\n          \"kind\": %s\n        }",
+				sbMeta.WriteString(fmt.Sprintf("\n        {\n          \"path\": %s,\n          \"kind\": %s\n        }",
 					js_printer.QuoteForJSON(c.res.PrettyPath(logger.Path{Text: importAbsPath, Namespace: "file"}), c.options.ASCIIOnly),
 					js_printer.QuoteForJSON(continueData.crossChunkImportRecords[i].Kind.StringForMetafile(), c.options.ASCIIOnly)))
 			}
 			if !isFirstMeta {
-				jMeta.AddString("\n      ")
+				sbMeta.WriteString("\n      ")
 			}
 
 			// Print exports
-			jMeta.AddString("],\n      \"exports\": [")
+			sbMeta.WriteString("],\n      \"exports\": [")
 			var aliases []string
 			if c.options.OutputFormat.KeepES6ImportExportSyntax() {
 				if chunk.isEntryPoint {
@@ -3831,19 +3831,19 @@ func (repr *chunkReprJS) generate(c *linkerContext, chunk *chunkInfo) func(gener
 				if isFirstMeta {
 					isFirstMeta = false
 				} else {
-					jMeta.AddString(",")
+					sbMeta.WriteString(",")
 				}
-				jMeta.AddString(fmt.Sprintf("\n        %s",
+				sbMeta.WriteString(fmt.Sprintf("\n        %s",
 					js_printer.QuoteForJSON(alias, c.options.ASCIIOnly)))
 			}
 			if !isFirstMeta {
-				jMeta.AddString("\n      ")
+				sbMeta.WriteString("\n      ")
 			}
 			if chunk.isEntryPoint {
 				entryPoint := c.files[chunk.sourceIndex].source.PrettyPath
-				jMeta.AddString(fmt.Sprintf("],\n      \"entryPoint\": %s,\n      \"inputs\": {", js_printer.QuoteForJSON(entryPoint, c.options.ASCIIOnly)))
+				sbMeta.WriteString(fmt.Sprintf("],\n      \"entryPoint\": %s,\n      \"inputs\": {", js_printer.QuoteForJSON(entryPoint, c.options.ASCIIOnly)))
 			} else {
-				jMeta.AddString("],\n      \"inputs\": {")
+				sbMeta.WriteString("],\n      \"inputs\": {")
 			}
 		}
 
@@ -3855,7 +3855,7 @@ func (repr *chunkReprJS) generate(c *linkerContext, chunk *chunkInfo) func(gener
 		var metaByteCount map[string]int
 		commentSet := make(map[string]bool)
 		prevComment := uint32(0)
-		if c.options.AbsMetadataFile != "" {
+		if c.options.NeedsMetafile {
 			metaOrder = make([]string, 0, len(compileResults))
 			metaByteCount = make(map[string]int, len(compileResults))
 		}
@@ -3919,7 +3919,7 @@ func (repr *chunkReprJS) generate(c *linkerContext, chunk *chunkInfo) func(gener
 				}
 
 				// Include this file in the metadata
-				if c.options.AbsMetadataFile != "" {
+				if c.options.NeedsMetafile {
 					// Accumulate file sizes since a given file may be split into multiple parts
 					path := c.files[compileResult.sourceIndex].source.PrettyPath
 					if count, ok := metaByteCount[path]; ok {
@@ -4001,10 +4001,10 @@ func (repr *chunkReprJS) generate(c *linkerContext, chunk *chunkInfo) func(gener
 			// Write the generated source map as an external file
 			if writeFile {
 				// Optionally add metadata about the file
-				var jsonMetadataChunk []byte
-				if c.options.AbsMetadataFile != "" {
-					jsonMetadataChunk = []byte(fmt.Sprintf(
-						"{\n      \"imports\": [],\n      \"exports\": [],\n      \"inputs\": {},\n      \"bytes\": %d\n    }", len(sourceMap)))
+				var jsonMetadataChunk string
+				if c.options.NeedsMetafile {
+					jsonMetadataChunk = fmt.Sprintf(
+						"{\n      \"imports\": [],\n      \"exports\": [],\n      \"inputs\": {},\n      \"bytes\": %d\n    }", len(sourceMap))
 				}
 
 				// Figure out the base name for the source map which may include the content hash
@@ -4063,23 +4063,23 @@ func (repr *chunkReprJS) generate(c *linkerContext, chunk *chunkInfo) func(gener
 		}
 
 		// End the metadata
-		var jsonMetadataChunk []byte
-		if c.options.AbsMetadataFile != "" {
+		var jsonMetadataChunk string
+		if c.options.NeedsMetafile {
 			isFirstMeta := true
 			for _, path := range metaOrder {
 				if isFirstMeta {
 					isFirstMeta = false
 				} else {
-					jMeta.AddString(",")
+					sbMeta.WriteString(",")
 				}
-				jMeta.AddString(fmt.Sprintf("\n        %s: {\n          \"bytesInOutput\": %d\n        }",
+				sbMeta.WriteString(fmt.Sprintf("\n        %s: {\n          \"bytesInOutput\": %d\n        }",
 					js_printer.QuoteForJSON(path, c.options.ASCIIOnly), metaByteCount[path]))
 			}
 			if !isFirstMeta {
-				jMeta.AddString("\n      ")
+				sbMeta.WriteString("\n      ")
 			}
-			jMeta.AddString(fmt.Sprintf("},\n      \"bytes\": %d\n    }", len(jsContents)))
-			jsonMetadataChunk = jMeta.Done()
+			sbMeta.WriteString(fmt.Sprintf("},\n      \"bytes\": %d\n    }", len(jsContents)))
+			jsonMetadataChunk = sbMeta.String()
 		}
 
 		results = append(results, OutputFile{
@@ -4226,22 +4226,22 @@ func (repr *chunkReprCSS) generate(c *linkerContext, chunk *chunkInfo) func(gene
 		}
 
 		// Start the metadata
-		jMeta := js_printer.Joiner{}
-		if c.options.AbsMetadataFile != "" {
+		sbMeta := strings.Builder{}
+		if c.options.NeedsMetafile {
 			isFirstMeta := true
-			jMeta.AddString("{\n      \"imports\": [")
+			sbMeta.WriteString("{\n      \"imports\": [")
 			for i, importAbsPath := range continueData.crossChunkAbsPaths {
 				if isFirstMeta {
 					isFirstMeta = false
 				} else {
-					jMeta.AddString(",")
+					sbMeta.WriteString(",")
 				}
-				jMeta.AddString(fmt.Sprintf("\n        {\n          \"path\": %s,\n          \"kind\": %s\n        }",
+				sbMeta.WriteString(fmt.Sprintf("\n        {\n          \"path\": %s,\n          \"kind\": %s\n        }",
 					js_printer.QuoteForJSON(c.res.PrettyPath(logger.Path{Text: importAbsPath, Namespace: "file"}), c.options.ASCIIOnly),
 					js_printer.QuoteForJSON(continueData.crossChunkImportRecords[i].Kind.StringForMetafile(), c.options.ASCIIOnly)))
 			}
 			if !isFirstMeta {
-				jMeta.AddString("\n      ")
+				sbMeta.WriteString("\n      ")
 			}
 			if chunk.isEntryPoint {
 				file := &c.files[chunk.sourceIndex]
@@ -4250,13 +4250,13 @@ func (repr *chunkReprCSS) generate(c *linkerContext, chunk *chunkInfo) func(gene
 				// importing CSS into JavaScript. We want this to be a 1:1 relationship
 				// and there is already an output file for the JavaScript entry point.
 				if _, ok := file.repr.(*reprCSS); ok {
-					jMeta.AddString(fmt.Sprintf("],\n      \"entryPoint\": %s,\n      \"inputs\": {",
+					sbMeta.WriteString(fmt.Sprintf("],\n      \"entryPoint\": %s,\n      \"inputs\": {",
 						js_printer.QuoteForJSON(file.source.PrettyPath, c.options.ASCIIOnly)))
 				} else {
-					jMeta.AddString("],\n      \"inputs\": {")
+					sbMeta.WriteString("],\n      \"inputs\": {")
 				}
 			} else {
-				jMeta.AddString("],\n      \"inputs\": {")
+				sbMeta.WriteString("],\n      \"inputs\": {")
 			}
 		}
 		isFirstMeta := true
@@ -4275,13 +4275,13 @@ func (repr *chunkReprCSS) generate(c *linkerContext, chunk *chunkInfo) func(gene
 			j.AddString(compileResult.printedCSS)
 
 			// Include this file in the metadata
-			if c.options.AbsMetadataFile != "" {
+			if c.options.NeedsMetafile {
 				if isFirstMeta {
 					isFirstMeta = false
 				} else {
-					jMeta.AddString(",")
+					sbMeta.WriteString(",")
 				}
-				jMeta.AddString(fmt.Sprintf("\n        %s: {\n          \"bytesInOutput\": %d\n        }",
+				sbMeta.WriteString(fmt.Sprintf("\n        %s: {\n          \"bytesInOutput\": %d\n        }",
 					js_printer.QuoteForJSON(c.files[compileResult.sourceIndex].source.PrettyPath, c.options.ASCIIOnly),
 					len(compileResult.printedCSS)))
 			}
@@ -4318,13 +4318,13 @@ func (repr *chunkReprCSS) generate(c *linkerContext, chunk *chunkInfo) func(gene
 		}
 
 		// End the metadata
-		var jsonMetadataChunk []byte
-		if c.options.AbsMetadataFile != "" {
+		var jsonMetadataChunk string
+		if c.options.NeedsMetafile {
 			if !isFirstMeta {
-				jMeta.AddString("\n      ")
+				sbMeta.WriteString("\n      ")
 			}
-			jMeta.AddString(fmt.Sprintf("},\n      \"bytes\": %d\n    }", len(cssContents)))
-			jsonMetadataChunk = jMeta.Done()
+			sbMeta.WriteString(fmt.Sprintf("},\n      \"bytes\": %d\n    }", len(cssContents)))
+			jsonMetadataChunk = sbMeta.String()
 		}
 
 		results = append(results, OutputFile{